# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "e-Demic API Documentation",
        "description": "API endpoints and descriptions for e-Demic Laravel API",
        "contact": {
            "name": "Developer : Sreejith B S",
            "email": "sreejith.bs@quinoid.in"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/auth/sendOtp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send SMS verification OTP code for users",
                "description": "API to send SMS verification OTP code for users. <br><br>Accepts `phone_number` and either `android_device_id` or `ios_device_id`.",
                "operationId": "sendOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "android_device_id": {
                                        "type": "string"
                                    },
                                    "ios_device_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone_number": "+918943406910",
                                    "android_device_id": "9774d56d682e549c",
                                    "ios_device_id": ""
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    }
                }
            }
        },
        "/auth/verifyOtp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify user entered OTP code",
                "description": "API to verify user entered OTP code. <br><br>Accepts `phone_number` and `code`. Returns `token` for user authorization purpose.",
                "operationId": "verifyOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "otp_code": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone_number": "+918943406910",
                                    "otp_code": 123456
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    }
                }
            }
        },
        "/messages/{index}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List all messages",
                "description": "API to list all messages, such as alerts, announcements etc on latest first basis. Default limit set as **10** messages per page. <br><br>Accepts `index` for lazy load messages.",
                "operationId": "messages",
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "description": "Page index to fetch messages",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}